Made changes suggested in review
