+++ docs/source/index.rst
@@ -88,7 +88,7 @@ In the directory organization above:
 
       * `2021 Google Android Derived Dataset <https://www.kaggle.com/c/google-smartphone-decimeter-challenge>`__
       * `2022 Google Android Derived Dataset <https://www.kaggle.com/competitions/smartphone-decimeter-2022>`__
-      * `Precise Ephemeris Dataset <https://cddis.nasa.gov/Data_and_Derived_Products/GNSS/gnss_mgex.html>`__
+      * `Precise Ephemeris Data <https://cddis.nasa.gov/Data_and_Derived_Products/GNSS/gnss_mgex.html>`__
 
   * The :code:`utils` directory contains utilities used to handle
     GNSS measurements, time conversions, visualizations, satellite
